Add files via upload
SQL and Python are two widely used technologies in the field of data analytics and data science. 



While SQL allows you to communicate with relational databases, Python helps you analyze and visualize data by using various libraries.



Exploring data with the combination of these tools can be a superpower. In this project, the connection into mySQL server was made using necessary Python libraries. Various functions and methods was established to create database, tables and manipulate tables.



In conclusion, a significant results was achieve with was converted to pandas data frame for future records.